refactor(T10131): decompose system.ts renderers into core/render/{session,orchestration,brain} (B6)
Migrates all 22 renderers (1302 LOC) from packages/cleo/src/cli/renderers/system.ts
to packages/core/src/render/ organized by family:

- session/ (11)       — briefing, blockers, next, current, doctor, session,
                        version, start, stop, schema, generic
- orchestration/ (4+) — tree, waves, plan, audit-reconstruct, stats
                        (+ renderCompletionBar helper)
- brain/ (6)          — maintenance, backfill, purge, plasticity-stats,
                        quality, export

Each renderer self-registers into the B5 renderer registry under
kind='generic' via a wrapLegacyRenderer adapter so future envelope-aware
callers can resolve them through renderEnvelopeForHuman(). The cleo
dispatcher in packages/cleo/src/cli/renderers/index.ts updates its imports
to pull these names from @cleocode/core instead of the deleted ./system.js.

Supporting helpers also migrated to core:
- colors.ts (BOLD/DIM/etc + statusColor/statusSymbol + priority helpers + BOX
  + hRule + shortDate). Old location kept as a re-export shim so status.ts
  and tasks.ts (B8 domain) continue to import from the historical path.
- tree-context.ts (TreeContext singleton). Old location kept as a re-export
  shim so commands/deps.ts continues to import from the historical path.
- cli-colorize.ts — ColorStyle->ANSI adapter extracted from system.ts
- format-label.ts — camelCase->Title Case helper extracted from system.ts
- legacy-adapter.ts — wrapLegacyRenderer bridging (data, quiet) -> Renderer<T>

Zero behavior change — all 197 renderer + audit-clioutput tests pass unchanged.
Resolves AGENTS.md Package-Boundary Check violation (rendering logic belongs
in core).

Subtasks T10147 (session), T10148 (orchestration), T10149 (brain) delivered.

Closes T10131
Subtasks: T10147, T10148, T10149
Epic: T10114
ADR: adr-077-human-render-contract
